+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 0.3.3 (2026-08-06)
 
 - Support stable and prerelease GitHub and Marketplace releases